requirements.txt
torch>=2.1.0
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
requests>=2.31.0
tqdm>=4.66.0
pyyaml>=6.0.1
scikit-learn>=1.3.0

